LayoutTests/fast/shadow-dom/cannot-hit-test-into-shadow-range-input-expected.txt

PASS range.commonAncestorContainer is rangeInputElement
PASS successfullyParsed is true

TEST COMPLETE
